Add a jobcount utility to detect the number of free jobs
If you can only decide e.g. how many threads to create, but not interact
with the job server for each task, this *could* help.

With many caveats though - as jobs start and finish this value varies,
and in "ideal" circumstances it should pretty much always be 0 since
there's plenty of work to go around.